# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface HeroContent {
  eyebrow: string;
  headline: string;
  subheadline: string;
  primaryCta: string;
  secondaryCta: string;
}

export interface Stat {
  value: string;
  label: string;
}

export interface Service {
  id: string;
  title: string;
  description: string;
  outcome: string;
  tags: string[];
}

export interface CaseStudy {
  id: string;
  client: string;
  industry: string;
  title: string;
  result: string;
  metric: string;
}

export interface Testimonial {
  quote: string;
  name: string;
  role: string;
  company: string;
}

export interface Faq {
  question: string;
  answer: string;
}

export interface BlogPost {
  id: string;
  category: string;
  title: string;
  excerpt: string;
  readTime: string;
}

export interface SiteContent {
  hero: HeroContent;
  intro: string;
  stats: Stat[];
  services: Service[];
  caseStudies: CaseStudy[];
  testimonials: Testimonial[];
  faqs: Faq[];
  posts: BlogPost[];
  techStack: string[];
  process: string[];
}

export type SiteContentInput = SiteContent;

export interface ContactRequestInput {
  /** @minLength 1 */
  name: string;
  email: string;
  company: string;
  service: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactRequest {
  id: number;
  name: string;
  email: string;
  company: string;
  service: string;
  message: string;
  status: string;
  createdAt: string;
}

export interface AdminOverview {
  contentSections: number;
  serviceCount: number;
  caseStudyCount: number;
  inquiryCount: number;
  latestInquiries: ContactRequest[];
}