# lib/api-zod/src/generated/types/adminOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactRequest } from './contactRequest';

export interface AdminOverview {
  contentSections: number;
  serviceCount: number;
  caseStudyCount: number;
  inquiryCount: number;
  latestInquiries: ContactRequest[];
}