# lib/api-zod/src/generated/types/contactRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactRequest {
  id: number;
  name: string;
  email: string;
  company: string;
  service: string;
  message: string;
  status: string;
  createdAt: string;
}