# lib/api-zod/src/generated/types/contactRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactRequestInput {
  /** @minLength 1 */
  name: string;
  email: string;
  company: string;
  service: string;
  /** @minLength 1 */
  message: string;
}