# lib/api-zod/src/generated/types/siteContent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPost } from './blogPost';
import type { CaseStudy } from './caseStudy';
import type { Faq } from './faq';
import type { HeroContent } from './heroContent';
import type { Service } from './service';
import type { Stat } from './stat';
import type { Testimonial } from './testimonial';

export interface SiteContent {
  hero: HeroContent;
  intro: string;
  stats: Stat[];
  services: Service[];
  caseStudies: CaseStudy[];
  testimonials: Testimonial[];
  faqs: Faq[];
  posts: BlogPost[];
  techStack: string[];
  process: string[];
}